requirements.txt
cachetools==5.3.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
urllib3==1.26.15
